requirements.txt
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.0.0
pywin32>=306
numpy>=1.24.0
pyperclip>=1.8.2
